Adiciona itens do desafio tentados, mas sem sucesso

+++ modulo2/apis-labenusers2/labenusers-com-components/src/components/screenUsersList.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import axios from 'axios'
 import styled from 'styled-components'
-import ScreenDetails from "./ScreenDetails";
+import ScreenDetails from "./screenDetails";
 
 const UserCard = styled.div`
 border: 1px solid black;
@@ -26,20 +26,19 @@ class ScreenUsersList extends React.Component {
         userId: ''
     }
 
-    changeScreen = () => {
-        switch (this.state.screen) {
-          case "userDetails":
-            return <ScreenDetails goToRegister={this.goToRegister} />;
-          default:
-            return <ScreenUsersList goToList={this.goToList} />;
+    changeScreen = (userId) => {
+        if (this.state.screen === 'userList') {
+          this.setState({screen: 'sreenDetails', userId: userId});
+        } else {
+            this.setState ({screen: 'userList', userId: ''})
         }
-      };
+    };
 
     componentDidMount() {
         this.getUsers()
     }
 
-    getUsers = async() => {
+    getUsers = () => {
         const urlUsers = 'https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users'
 
         axios.get(urlUsers, {
@@ -51,7 +50,7 @@ class ScreenUsersList extends React.Component {
         }).catch((err) => {
             alert('Ocorreu um problema, tente novamente: ', err)
         })
-
+/* 
         try {
             const res = await axios.get(urlUsers, {
                 headers: {
@@ -61,48 +60,53 @@ class ScreenUsersList extends React.Component {
             this.setState({ usersList: res.data })
         } catch (err) {
             alert('Ocorreu um problema, tente novamente: ', err)
-        }
+        } */
 
     }
 
     deletUser = (id) => {
+        if (window.confirm('Esta certo disso? Ler com voz de Silvio Santos')){
         const urlDelete = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`
+        
         axios.delete(urlDelete, {
             headers: {
                 Authorization: 'jil-moutinho-ailton'
             }
-        }).then((res) => {
+        }).then(res => {
             alert('Usuário apagado com sucesso!')
             {this.getUsers()}
-        }).catch((err) => {
-            alert('Ocorreu um problema, tente novamente: ', err)
         })
+        .catch(err => {
+            alert('Ocorreu um problema, tente novamente: ', err)
+        })} else {
+            alert('Que bom que não apagou o usuário!')
+        }
     }
 
-/*     handleFindName = (event) => {
-        this.setState({ FindName: event.target.value });
-      };
+    handleFindName = (event) => {
+        this.setState({ findName: event.target.value });
+    };
 
-    findUser = (FindName) => {
-        const urlGetByName = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/search?${FindName}`
-
+    findUser = () => {
+        const urlGetByName = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/search?name=${this.state.findName}&email=`;
+        
         axios.get(urlGetByName, {
             headers: {
                 Authorization: 'jil-moutinho-ailton'
             }
-        }).then((res) => {
-            
+        }).then(res => {
+            this.setState({ userList: res.data})
        console.log(res.data);
             
         }).catch((err) => {
             alert('Ocorreu um problema, tente novamente: ', err)
         })
-    } */
+    }
 
     render() {
         const usersListScreen = this.state.usersList.map((user) => {
             return (
-                <UserCard key={user.id}>
+                <UserCard onClick={() => this.changeScreen(user.id)} key={user.id}>
                     {user.name}
                     <ButtonDelete onClick={() => this.deletUser(user.id)}>Deletar usuário</ButtonDelete>
                 </UserCard>
@@ -111,17 +115,24 @@ class ScreenUsersList extends React.Component {
 
         return (
             <div>
-                <h2>Lista de usuários cadastrados</h2>
-                {usersListScreen}
-                <p>Procurando alguém específico?</p>
-{/*                 <input
-                    placeholder="Nome"
-                    value={this.state.FindName}
-                    onChange={this.handleEmail}
-                />
-                <button onClick={this.findUser}>Procurar</button> */}
-
-                <button onClick={this.props.goToRegister}>Ir para cadastro de usuários</button>
+                {this.state.screen === 'userList' ? (
+                    <div>
+                    <h2>Lista de usuários cadastrados</h2>
+                    {this.state.usersList.length === 0 && <div>Ainda crregando...</div>}
+                    {usersListScreen}
+                    <h3>Procurando alguém específico?</h3>
+                    <input
+                        placeholder="Nome"
+                        type='text"'
+                        value={this.state.findName}
+                        onChange={this.handleFindName}
+                    />
+                    <button onClick={this.findUser}>Procurar</button>
+                    <button onClick={this.props.goToRegister}>Ir para cadastro de usuários</button>
+                </div>
+                ) : (
+                    <ScreenDetails userId={this.state.userId} changeScreen={this.changeScreen} />
+                )}
             </div>
         )
     }
